requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
alembic==1.15.2
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
colour-science==0.4.6
comfyui_frontend_package==1.20.7
comfyui_workflow_templates==0.1.23
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
Deprecated==1.2.18
docopt==0.6.2
easydict==1.13
einops==0.8.1
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
gdown==5.2.0
gguf==0.17.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.31.4
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
macholib==1.16.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matrix-client==0.4.0
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numba==0.61.2
numpy<2.3
omegaconf==2.3.0
openai-whisper==20240930
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optuna==4.3.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pixeloe==0.1.4
platformdirs==4.3.8
pooch==1.8.2
primePy==1.3
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
PyMatting==1.1.14
PyNaCl==1.5.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
setuptools==80.7.1
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
spandrel==0.4.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
stringzilla==3.12.5
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tifffile==2025.5.26
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.3.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.3.1
torchmetrics==1.7.1
torchsde==0.2.6
torchvision==0.18.1
tqdm==4.67.1
trampoline==0.1.2
transformers==4.52.4
transparent-background==1.3.4
ttkthemes==3.2.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uv==0.7.9
wget==3.2
wrapt==1.17.2
yarl==1.20.0
PySide6
